+++ src/pages/messages/layout.jsx
@@ -6,10 +6,15 @@ import { useSearchParams } from "react-router";
 import { CircleTitleSubtitleSkeleton } from "../../components/admin/feeds/TopServiceSuggestions";
 import HeadingText from "../../components/HeadingText";
 import LightParagraph from "../../components/ParagraphText";
-import SEO from "../../components/SEO";
+import SEO, { createSEO } from "../../components/SEO";
 import { useAuth } from "../../context/userContext";
 
 import MessagesPage from "../messages";
+
+export const meta = () =>
+  createSEO({
+    title: "Messaging in connectize",
+  });
 
 export default function MessagesLayout() {
   const [searchParams] = useSearchParams();