+++ vite.config.mjs
@@ -1,11 +1,12 @@
 import { visualizer } from "rollup-plugin-visualizer";
-import react from "@vitejs/plugin-react";
+// import react from "@vitejs/plugin-react";
 import { reactRouter } from "@react-router/dev/vite";
 import { defineConfig } from "vite";
 
 export default defineConfig({
   plugins: [
     reactRouter(),
+
     // react(),
     // visualizer({
     //   open: true,