+++ changes.md
@@ -3,3 +3,7 @@
 
 -
 - Completed the migration to react-router-v7
+
+- merged all changes from react-router-v7 mirgration to the current state of main.
+- Updated all <SEO /> call to the react router v7 pattern
+- added <Links /> to the root layout to load css during build time